+++ src/polymorph/__init__.py
@@ -15,4 +15,4 @@ __all__ = [
     "SchemaDescriptor",
     "Sensitivity",
 ]
-__version__ = "0.4.0a1"
+__version__ = "0.4.0a2"